Refactor: Extract noqa methods and tests. (#5010)
