+++ src/sqlfluff/core/linter/linter.py
@@ -1,6 +1,5 @@
 """Defines the linter class."""
 
-import fnmatch
 import os
 import time
 import logging
@@ -14,7 +13,6 @@ from typing import (
     Set,
     Tuple,
     Type,
-    Dict,
     cast,
 )
 
@@ -30,7 +28,7 @@ from sqlfluff.core.errors import (
     SQLFluffSkipFile,
     SQLFluffUserError,
 )
-from sqlfluff.core.parser import Lexer, Parser, RegexLexer
+from sqlfluff.core.parser import Lexer, Parser
 from sqlfluff.core.file_helpers import get_encoding
 from sqlfluff.core.templaters import TemplatedFile
 from sqlfluff.core.rules import get_ruleset
@@ -39,15 +37,14 @@ from sqlfluff.core.config import FluffConfig, ConfigLoader, progress_bar_configu
 # Classes needed only for type checking
 from sqlfluff.core.parser.segments.base import BaseSegment, SourceFix
 from sqlfluff.core.parser.segments.meta import MetaSegment
-from sqlfluff.core.parser.segments.raw import RawSegment
 from sqlfluff.core.rules import BaseRule, RulePack
 
 from sqlfluff.core.linter.common import (
     RuleTuple,
     ParsedString,
-    NoQaDirective,
     RenderedFile,
 )
+from sqlfluff.core.linter.noqa import IgnoreMask
 from sqlfluff.core.linter.linted_file import (
     LintedFile,
     FileTimings,
@@ -265,83 +262,6 @@ class Linter:
         return parsed, violations
 
     @staticmethod
-    def parse_noqa(
-        comment: str,
-        line_no: int,
-        reference_map: Dict[str, Set[str]],
-    ):
-        """Extract ignore mask entries from a comment string."""
-        # Also trim any whitespace afterward
-
-        # Comment lines can also have noqa e.g.
-        # --dafhsdkfwdiruweksdkjdaffldfsdlfjksd -- noqa: LT05
-        # Therefore extract last possible inline ignore.
-        comment = [c.strip() for c in comment.split("--")][-1]
-
-        if comment.startswith("noqa"):
-            # This is an ignore identifier
-            comment_remainder = comment[4:]
-            if comment_remainder:
-                if not comment_remainder.startswith(":"):
-                    return SQLParseError(
-                        "Malformed 'noqa' section. Expected 'noqa: <rule>[,...]",
-                        line_no=line_no,
-                    )
-                comment_remainder = comment_remainder[1:].strip()
-                if comment_remainder:
-                    action: Optional[str]
-                    if "=" in comment_remainder:
-                        action, rule_part = comment_remainder.split("=", 1)
-                        if action not in {"disable", "enable"}:  # pragma: no cover
-                            return SQLParseError(
-                                "Malformed 'noqa' section. "
-                                "Expected 'noqa: enable=<rule>[,...] | all' "
-                                "or 'noqa: disable=<rule>[,...] | all",
-                                line_no=line_no,
-                            )
-                    else:
-                        action = None
-                        rule_part = comment_remainder
-                        if rule_part in {"disable", "enable"}:
-                            return SQLParseError(
-                                "Malformed 'noqa' section. "
-                                "Expected 'noqa: enable=<rule>[,...] | all' "
-                                "or 'noqa: disable=<rule>[,...] | all",
-                                line_no=line_no,
-                            )
-                    rules: Optional[Tuple[str, ...]]
-                    if rule_part != "all":
-                        # Rules can be globs therefore we compare to the rule_set to
-                        # expand the globs.
-                        unexpanded_rules = tuple(
-                            r.strip() for r in rule_part.split(",")
-                        )
-                        # We use a set to do natural deduplication.
-                        expanded_rules: Set[str] = set()
-                        for r in unexpanded_rules:
-                            matched = False
-                            for expanded in (
-                                reference_map[x]
-                                for x in fnmatch.filter(reference_map.keys(), r)
-                            ):
-                                expanded_rules |= expanded
-                                matched = True
-
-                            if not matched:
-                                # We were unable to expand the glob.
-                                # Therefore assume the user is referencing
-                                # a special error type (e.g. PRS, LXR, or TMP)
-                                # and add this to the list of rules to ignore.
-                                expanded_rules.add(r)
-                        # Sort for consistency
-                        rules = tuple(sorted(expanded_rules))
-                    else:
-                        rules = None
-                    return NoQaDirective(line_no, rules, action)
-            return NoQaDirective(line_no, None, None)
-        return None
-
-    @staticmethod
     def remove_templated_errors(
         linting_errors: List[SQLBaseError],
     ) -> List[SQLBaseError]:
@@ -428,69 +348,6 @@ class Linter:
         )
 
     @classmethod
-    def extract_ignore_from_comment(
-        cls,
-        comment: RawSegment,
-        reference_map: Dict[str, Set[str]],
-    ):
-        """Extract ignore mask entries from a comment segment."""
-        # Also trim any whitespace afterward
-        comment_content = comment.raw_trimmed().strip()
-        comment_line, _ = comment.pos_marker.source_position()
-        result = cls.parse_noqa(comment_content, comment_line, reference_map)
-        if isinstance(result, SQLParseError):
-            result.segment = comment
-        return result
-
-    @classmethod
-    def extract_ignore_mask_tree(
-        cls,
-        tree: BaseSegment,
-        reference_map: Dict[str, Set[str]],
-    ) -> Tuple[List[NoQaDirective], List[SQLBaseError]]:
-        """Look for inline ignore comments and return NoQaDirectives."""
-        ignore_buff: List[NoQaDirective] = []
-        violations: List[SQLBaseError] = []
-        for comment in tree.recursive_crawl("comment"):
-            if comment.is_type("inline_comment"):
-                ignore_entry = cls.extract_ignore_from_comment(comment, reference_map)
-                if isinstance(ignore_entry, SQLParseError):
-                    violations.append(ignore_entry)
-                elif ignore_entry:
-                    ignore_buff.append(ignore_entry)
-        if ignore_buff:
-            linter_logger.info("Parsed noqa directives from file: %r", ignore_buff)
-        return ignore_buff, violations
-
-    @classmethod
-    def extract_ignore_mask_source(
-        cls,
-        source: str,
-        inline_comment_regex: RegexLexer,
-        reference_map: Dict[str, Set[str]],
-    ) -> Tuple[List[NoQaDirective], List[SQLBaseError]]:
-        """Look for inline ignore comments and return NoQaDirectives.
-
-        Very similar to extract_ignore_mask_tree(), but can be run on raw source
-        (i.e. does not require the code to have parsed successfully).
-        """
-        ignore_buff: List[NoQaDirective] = []
-        violations: List[SQLBaseError] = []
-        for idx, line in enumerate(source.split("\n")):
-            match = inline_comment_regex.search(line) if line else None
-            if match:
-                ignore_entry = cls.parse_noqa(
-                    line[match[0] : match[1]], idx + 1, reference_map
-                )
-                if isinstance(ignore_entry, SQLParseError):
-                    violations.append(ignore_entry)  # pragma: no cover
-                elif ignore_entry:
-                    ignore_buff.append(ignore_entry)
-        if ignore_buff:
-            linter_logger.info("Parsed noqa directives from file: %r", ignore_buff)
-        return ignore_buff, violations
-
-    @classmethod
     def lint_fix_parsed(
         cls,
         tree: BaseSegment,
@@ -500,7 +357,7 @@ class Linter:
         fname: Optional[str] = None,
         templated_file: Optional[TemplatedFile] = None,
         formatter: Any = None,
-    ) -> Tuple[BaseSegment, List[SQLBaseError], List[NoQaDirective], RuleTimingsType]:
+    ) -> Tuple[BaseSegment, List[SQLBaseError], Optional[IgnoreMask], RuleTimingsType]:
         """Lint and optionally fix a tree object."""
         # Keep track of the linting errors on the very first linter pass. The
         # list of issues output by "lint" and "fix" only includes issues present
@@ -524,12 +381,10 @@ class Linter:
 
         # Look for comment segments which might indicate lines to ignore.
         if not config.get("disable_noqa"):
-            ignore_buff, ivs = cls.extract_ignore_mask_tree(
-                tree, rule_pack.reference_map
-            )
+            ignore_mask, ivs = IgnoreMask.from_tree(tree, rule_pack.reference_map)
             initial_linting_errors += ivs
         else:
-            ignore_buff = []
+            ignore_mask = None
 
         save_tree = tree
         # There are two phases of rule running.
@@ -599,7 +454,7 @@ class Linter:
                         dialect=config.get("dialect_obj"),
                         fix=fix,
                         templated_file=templated_file,
-                        ignore_mask=ignore_buff,
+                        ignore_mask=ignore_mask,
                         fname=fname,
                         config=config,
                     )
@@ -700,7 +555,7 @@ class Linter:
                     # Reason: When the linter hits the loop limit, the file is often
                     # messy, e.g. some of the fixes were applied repeatedly, possibly
                     # other weird things. We don't want the user to see this junk!
-                    return save_tree, initial_linting_errors, ignore_buff, rule_timings
+                    return save_tree, initial_linting_errors, ignore_mask, rule_timings
 
         if config.get("ignore_templated_areas", default=True):
             initial_linting_errors = cls.remove_templated_errors(initial_linting_errors)
@@ -708,7 +563,7 @@ class Linter:
         linter_logger.info("\n###\n#\n# {}\n#\n###".format("Fixed Tree:"))
         linter_logger.info("\n" + tree.stringify())
 
-        return tree, initial_linting_errors, ignore_buff, rule_timings
+        return tree, initial_linting_errors, ignore_mask, rule_timings
 
     @classmethod
     def lint_parsed(
@@ -729,7 +584,7 @@ class Linter:
             (
                 tree,
                 initial_linting_errors,
-                ignore_buff,
+                ignore_mask,
                 rule_timings,
             ) = cls.lint_fix_parsed(
                 parsed.tree,
@@ -749,14 +604,14 @@ class Linter:
         else:
             # If no parsed tree, set to None
             tree = None
-            ignore_buff = []
+            ignore_mask = None
             rule_timings = []
             if not parsed.config.get("disable_noqa"):
                 # Templating and/or parsing have failed. Look for "noqa"
                 # comments (the normal path for identifying these comments
                 # requires access to the parse tree, and because of the failure,
                 # we don't have a parse tree).
-                ignore_buff, ignore_violations = cls.extract_ignore_mask_source(
+                ignore_mask, ignore_violations = IgnoreMask.from_source(
                     parsed.source_str,
                     [
                         lm
@@ -778,7 +633,7 @@ class Linter:
             LintedFile.deduplicate_in_source_space(violations),
             FileTimings(time_dict, rule_timings),
             tree,
-            ignore_mask=ignore_buff,
+            ignore_mask=ignore_mask,
             templated_file=parsed.templated_file,
             encoding=encoding,
         )
