+++ test/core/linter/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for sqlfluff.core.linter."""
